Rebasing front-end parts.
