+++ app/javascript/mastodon/features/custom_filters_modal/custom_filter_types.ts
@@ -1,0 +1,4 @@
+export enum CustomFilterTypes {
+  Account = 'account',
+  Status = 'status',
+}
