+++ app/javascript/mastodon/features/custom_filters_modal/new_filter.jsx
